refactor: Constants 위치 변경
