+++ src/main/kotlin/racingcar/domain/Constants.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object Constants {
     // Game
